+++ core/src/main/java/org/seedstack/seed/core/internal/init/BaseConfiguration.java
@@ -8,10 +8,11 @@
 package org.seedstack.seed.core.internal.init;
 
 import org.seedstack.coffig.Coffig;
-import org.seedstack.coffig.provider.EnvironmentVariableProvider;
+import org.seedstack.coffig.provider.EnvironmentProvider;
 import org.seedstack.coffig.provider.JacksonProvider;
 import org.seedstack.coffig.provider.PrefixProvider;
-import org.seedstack.coffig.provider.SystemPropertyProvider;
+import org.seedstack.coffig.provider.PropertiesProvider;
+import org.seedstack.coffig.provider.SystemPropertiesProvider;
 import org.seedstack.seed.core.Seed;
 import org.seedstack.seed.core.internal.configuration.ConfigurationPriority;
 import org.seedstack.seed.core.internal.configuration.PrioritizedProvider;
@@ -23,15 +24,13 @@ import javax.validation.ValidatorFactory;
 import java.io.IOException;
 import java.net.URL;
 import java.util.Enumeration;
+import java.util.HashSet;
+import java.util.Set;
 
 public class BaseConfiguration {
     private static final Logger LOGGER = LoggerFactory.getLogger(BaseConfiguration.class);
     private static final String ENVIRONMENT_VARIABLES_PREFIX = "env";
     private static final String SYSTEM_PROPERTIES_PREFIX = "sys";
-    private static final String BASE_CONFIGURATION_PROVIDER = "base-config";
-    private static final String BASE_OVERRIDE_CONFIGURATION_PROVIDER = "base-config-override";
-    private static final String ENV_CONFIGURATION_PROVIDER = "environment-variables";
-    private static final String SYS_CONFIGURATION_PROVIDER = "system-properties";
     private final Coffig baseConfiguration;
 
     private static class Holder {
@@ -46,23 +45,27 @@ public class BaseConfiguration {
         baseConfiguration = Coffig.builder()
                 .withProviders(new PrioritizedProvider()
                         .registerProvider(
-                                BASE_CONFIGURATION_PROVIDER,
                                 buildJacksonProvider("application.yaml", "application.yml", "application.json"),
                                 ConfigurationPriority.BASE
                         )
                         .registerProvider(
-                                BASE_OVERRIDE_CONFIGURATION_PROVIDER,
+                                buildPropertiesProvider("application.properties"),
+                                ConfigurationPriority.BASE
+                        )
+                        .registerProvider(
                                 buildJacksonProvider("application.override.yaml", "application.override.yml", "application.override.json"),
                                 ConfigurationPriority.BASE_OVERRIDE
                         )
                         .registerProvider(
-                                ENV_CONFIGURATION_PROVIDER,
-                                new PrefixProvider<>(ENVIRONMENT_VARIABLES_PREFIX, new EnvironmentVariableProvider()),
+                                buildPropertiesProvider("application.override.properties"),
+                                ConfigurationPriority.BASE_OVERRIDE
+                        )
+                        .registerProvider(
+                                new PrefixProvider<>(ENVIRONMENT_VARIABLES_PREFIX, new EnvironmentProvider()),
                                 ConfigurationPriority.ENVIRONMENT_VARIABLES
                         )
                         .registerProvider(
-                                SYS_CONFIGURATION_PROVIDER,
-                                new PrefixProvider<>(SYSTEM_PROPERTIES_PREFIX, new SystemPropertyProvider()),
+                                new PrefixProvider<>(SYSTEM_PROPERTIES_PREFIX, new SystemPropertiesProvider()),
                                 ConfigurationPriority.SYSTEM_PROPERTIES
                         ))
                 .enableValidation(validatorFactory)
@@ -71,16 +74,28 @@ public class BaseConfiguration {
 
     private JacksonProvider buildJacksonProvider(String... resourceNames) {
         JacksonProvider jacksonProvider = new JacksonProvider();
+        getResources(resourceNames).forEach(jacksonProvider::addSource);
+        return jacksonProvider;
+    }
+
+    private PropertiesProvider buildPropertiesProvider(String... resourceNames) {
+        PropertiesProvider propertiesProvider = new PropertiesProvider();
+        getResources(resourceNames).forEach(propertiesProvider::addSource);
+        return propertiesProvider;
+    }
+
+    private Set<URL> getResources(String... resourceNames) {
+        Set<URL> result = new HashSet<>();
         for (String resourceName : resourceNames) {
             try {
                 Enumeration<URL> configResources = ClassLoaders.findMostCompleteClassLoader(Seed.class).getResources(resourceName);
                 while (configResources.hasMoreElements()) {
-                    jacksonProvider.addSource(configResources.nextElement());
+                    result.add(configResources.nextElement());
                 }
             } catch (IOException e) {
-                LOGGER.warn("I/O error during detection of configuration files", e);
+                LOGGER.warn("I/O error during detection of configuration file " + resourceName, e);
             }
         }
-        return jacksonProvider;
+        return result;
     }
 }